uplink: add `panic` extern
